+++ src/features/todos/TodoForm.tsx
@@ -9,8 +9,8 @@ import {
   Button,
 } from "@mui/material";
 import { useNavigate, useParams } from "react-router-dom";
-import { useAppDispatch, useAppSelector } from "../../hooks";
-import { addTodo, editTodo, fetchTodos } from "./todosSlice";
+import { useAppDispatch, useAppSelector } from "../../store";
+import { addTodo, editTodo, fetchTodos } from "../../store/TODO/todosSlice";
 import { unwrapResult } from "@reduxjs/toolkit";
 
 import { useForm, Controller } from "react-hook-form";
