+++ src/main.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./app/store";
+import { store } from "./store";
 import { BrowserRouter } from "react-router-dom";
 import { App } from "./App";
