+++ src/store/TODO/todosSlice.ts
@@ -3,7 +3,7 @@ import {
   createAsyncThunk,
   type PayloadAction,
 } from "@reduxjs/toolkit";
-import * as service from "../../services/todoService";
+import API from "../../services/axiosInstance";
 import type { Todo } from "../../types/Task/types";
 
 interface TodosState {
@@ -18,28 +18,34 @@ const initialState: TodosState = {
   error: null,
 };
 
+const getTodos = () => API.get<Todo[]>("/");
+const createTodo = (todo: Omit<Todo, "id" | "createdAt">) =>
+  API.post<Todo>("/", todo);
+const updateTodo = (todo: Todo) => API.put<Todo>(`/${todo.id}`, todo);
+const deleteTodo = (id: string) => API.delete(`/${id}`);
+
 export const fetchTodos = createAsyncThunk("todos/fetch", async () => {
-  const res = await service.getTodos();
-  return res.data as Todo[];
+  const res = await getTodos();
+  return res.data;
 });
 
 export const addTodo = createAsyncThunk(
   "todos/add",
   async (todo: Omit<Todo, "id" | "createdAt">) => {
-    const res = await service.createTodo(todo);
-    return res.data as Todo;
+    const res = await createTodo(todo);
+    return res.data;
   }
 );
 
 export const editTodo = createAsyncThunk("todos/edit", async (todo: Todo) => {
-  const res = await service.updateTodo(todo);
-  return res.data as Todo;
+  const res = await updateTodo(todo);
+  return res.data;
 });
 
 export const removeTodo = createAsyncThunk(
   "todos/remove",
   async (id: string) => {
-    await service.deleteTodo(id);
+    await deleteTodo(id);
     return id;
   }
 );
@@ -50,7 +56,6 @@ const todosSlice = createSlice({
   reducers: {},
   extraReducers: (builder) => {
     builder
-
       .addCase(fetchTodos.pending, (state) => {
         state.loading = true;
         state.error = null;
@@ -63,11 +68,9 @@ const todosSlice = createSlice({
         state.loading = false;
         state.error = action.error.message ?? "Failed to fetch todos";
       })
-
       .addCase(addTodo.fulfilled, (state, action: PayloadAction<Todo>) => {
         state.items.push(action.payload);
       })
-
       .addCase(editTodo.fulfilled, (state, action: PayloadAction<Todo>) => {
         const ix = state.items.findIndex(
           (todo) => todo.id === action.payload.id
@@ -76,7 +79,6 @@ const todosSlice = createSlice({
           state.items[ix] = action.payload;
         }
       })
-
       .addCase(removeTodo.fulfilled, (state, action: PayloadAction<string>) => {
         state.items = state.items.filter((todo) => todo.id !== action.payload);
       });
