+++ src/store/index.ts
@@ -1,6 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
-import todosReducer from "../features/todos/todosSlice";
+import todosReducer from "../store/TODO/todosSlice";
 import themeReducer from "../features/theme/themeSlice";
+
+import {
+  type TypedUseSelectorHook,
+  useDispatch,
+  useSelector,
+} from "react-redux";
 
 export const store = configureStore({
   reducer: {
@@ -11,3 +17,6 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
